Add files via upload
This is a COVID-19 probability website. Using Jupyter notebook and Logistic Regression and Double Variation in accordance with Machine learning. I have used flask, pickle to dump the data and used CSV files to hold the random data. Using flask I have created to HTML files and integrated it with BootStrap. Hope you take use of this project.